refactor: split the project into workspaces
