+++ flow_api/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "flow_api"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = { version = "1.28.0", features = ["full"] }
+axum = "0.6"
+tower-http = { version = "0.4", features = ["fs"] }
