+++ flow_data/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "flow_data"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rusqlite = { version = "0.28.0", features = ["bundled"] }
+chrono = "0.4.24"
+directories = "5.0"
