+++ flow_data/src/lib.rs
@@ -1,0 +1,11 @@
+mod directory;
+
+pub mod app_data;
+pub mod data;
+pub mod db;
+pub mod project;
+pub mod task;
+
+pub use directory::Directory;
+pub mod utils;
+pub use rusqlite::Error;
