+++ src/main.rs
@@ -1,10 +1,7 @@
 use cli::App;
 
-mod api;
-mod app_data;
 mod cli;
 mod constants;
-mod db;
 mod project;
 mod task;
 mod utils;
