--- src/utils/test_utils.rs
@@ -1,7 +1,0 @@
-pub struct After<C: FnMut()>(pub C);
-
-impl<C: FnMut()> Drop for After<C> {
-    fn drop(&mut self) {
-        (self.0)()
-    }
-}
